backend/requirements.txt
django==4.1.12
djangorestframework==3.14.0
markdown==3.4.1
psycopg2==2.9.5
uvicorn==0.21.1
gunicorn==20.1.0
django-cors-headers==3.14
tzdata==2023.3
pymongo==3.12.3
djongo==1.3.6
requests==2.31.0
drf-yasg==1.21.7
